refactor: navigation color changer
Set current background under the nav items dinamically based on their
position and the hero bounding rectangle. Ideally this would be generic
enough to allow every component with a dark background to inform their
bounding rectangles, but this would be overkill for such a small project
with only one component with dark background.