fix: fix payment link for custom gift amount
When the user selected a custom amount to gift, the payment link
wouldn't work because it tried to look for the price of the item in the
products list, but since it's a custom item it wouldn't find. Fix this
by making the payment link generator look first for custom values, then
look for the item price in the products list.